🐛 Fix landing hero demo hydration mismatch
The hero demo called getDemoDays(new Date()) inside a "use cache" scope
(the pages are "use cache" with cacheLife("days"), and cacheComponents is
on). That baked the dates into the prerendered HTML and RSC payload at
build time, and since those are separate cache entries that revalidate
independently, one could be regenerated across the weekly rollover while
the other was not — leaving the served HTML and the hydration payload
disagreeing.

The date set flips at UTC Friday 00:00, so the mismatch was intermittent
and self-healing, which is why it read as a random hydration failure on
initial load.

Derive an anchor (the first Thursday at least a week out) as a plain
YYYY-MM-DD string and key the cached demo on it. The HTML and the RSC
payload are now produced from the same key, so they cannot straddle the
rollover. cacheLife("hours") keeps the rendered dates from trailing the
rollover for long.

Also convert MobileDemo from one large "use client" component into a
server component, with the interactive parts split into small islands
(VoteSelector, VoteButton). This keeps the demo server rendered and
cacheable while preserving the voting interaction.
