+++ apps/landing/src/components/home/hero-demo/demo-data.ts
@@ -35,21 +35,28 @@ const HOUR_MS = 60 * 60 * 1000;
 const DAY_MS = 24 * HOUR_MS;
 const SLOT_HOURS = [12, 18];
 
-// The demo dates are floating wall times constructed in UTC and always
-// formatted with timeZone: "UTC", so the rendered clock times don't depend
-// on the server's time zone.
-export function getDemoDays(now: Date): DemoDay[] {
+// The first Thursday at least a week out, as a plain YYYY-MM-DD string. This is
+// the only thing about the demo that depends on the current time, so it is the
+// cache key for the rendered demo: everything below is a pure function of it.
+// Keying on the value rather than on a duration means the HTML and the RSC
+// payload can never be generated on opposite sides of a rollover, which would
+// otherwise surface as a hydration mismatch.
+export function getDemoAnchor(now: Date) {
   let day = new Date(now.getTime() + 7 * DAY_MS);
   while (day.getUTCDay() !== 4) {
     day = new Date(day.getTime() + DAY_MS);
   }
+  return day.toISOString().slice(0, 10);
+}
+
+// The demo dates are floating wall times constructed in UTC and always
+// formatted with timeZone: "UTC", so the rendered clock times don't depend
+// on the server's time zone.
+export function getDemoDays(anchor: string): DemoDay[] {
+  const [year, month, day] = anchor.split("-").map(Number);
+  const firstThursday = Date.UTC(year, month - 1, day);
   return [0, 1, 2, 3].map((week) => {
-    const thursday = new Date(day.getTime() + week * 7 * DAY_MS);
-    const midnight = Date.UTC(
-      thursday.getUTCFullYear(),
-      thursday.getUTCMonth(),
-      thursday.getUTCDate(),
-    );
+    const midnight = firstThursday + week * 7 * DAY_MS;
     return {
       date: new Date(midnight),
       options: SLOT_HOURS.map((hour) => ({
