+++ apps/landing/src/components/home/hero-demo/hero-demo.tsx
@@ -1,37 +1,36 @@
+import { cacheLife } from "next/cache";
 import { getTranslation } from "@/i18n/server";
-import { formatDemoParts, getDemoDays, getScores } from "./demo-data";
+import { getDemoAnchor, getDemoDays, getScores } from "./demo-data";
 import { DesktopDemo } from "./desktop-demo";
 import { MobileDemo } from "./mobile-demo";
 import { TryItPrompt } from "./try-it-prompt";
 
 // A presentational replica of the poll screens, deliberately decoupled from
-// apps/web so the landing page can't be broken by app changes. Dates are
-// computed at render time and cached with the page (cacheLife on the pages
-// that use this), so they refresh with the page cache instead of going stale
-// like a screenshot.
+// apps/web so the landing page can't be broken by app changes.
 //
-// The desktop shot is decoration (aria-hidden, no interactive elements); the
-// phone is a working demo, so date and time strings are pre-formatted here on
-// the server and passed down as plain props — client-side Intl could disagree
-// with the server and cause hydration mismatches.
+// The dates are derived from an anchor — the first Thursday at least a week
+// out — and the whole demo is a pure function of it. Deriving the anchor first
+// and passing it in as the cache key is what keeps hydration deterministic:
+// the HTML and the RSC payload are produced from the same key, so they cannot
+// be generated on opposite sides of the weekly rollover and disagree. Revalidating
+// hourly keeps the rendered dates from trailing the rollover for long.
 export const HeroDemo = async ({ locale }: { locale: string }) => {
-  const { t } = await getTranslation<"home">(locale, "home");
-  const days = getDemoDays(new Date());
-  const scores = getScores(days);
-  const format = formatDemoParts(locale);
+  "use cache";
+  cacheLife("hours");
+  return <CachedDemo locale={locale} anchor={getDemoAnchor(new Date())} />;
+};
 
-  let optionIndex = -1;
-  const mobileDays = days.map((day) => ({
-    label: format.fullDay.format(day.date),
-    options: day.options.map((option) => {
-      optionIndex += 1;
-      return {
-        id: option.start.toISOString(),
-        time: format.time.formatRange(option.start, option.end),
-        score: scores[optionIndex],
-      };
-    }),
-  }));
+const CachedDemo = async ({
+  locale,
+  anchor,
+}: {
+  locale: string;
+  anchor: string;
+}) => {
+  "use cache";
+  const { t } = await getTranslation<"home">(locale, "home");
+  const days = getDemoDays(anchor);
+  const scores = getScores(days);
 
   return (
     <div className="relative z-10 mb-12 md:mx-auto md:w-fit md:max-w-full">
@@ -50,7 +49,7 @@ export const HeroDemo = async ({ locale }: { locale: string }) => {
             defaultValue: "Go ahead, try voting!",
           })}
         />
-        <MobileDemo days={mobileDays} />
+        <MobileDemo locale={locale} days={days} scores={scores} t={t} />
       </div>
     </div>
   );
