+++ apps/landing/src/components/home/hero-demo/mobile-demo.tsx
@@ -1,91 +1,30 @@
-"use client";
-import { posthog } from "@rallly/posthog/client";
-import { cn } from "@rallly/ui";
-import { CircleCheckIcon, UserIcon, XIcon } from "lucide-react";
-import * as m from "motion/react-m";
-import Link from "next/link";
-import * as React from "react";
-import { Trans } from "@/i18n/client/trans";
-import { useTranslation } from "@/i18n/client/use-translation";
-import { linkToApp } from "@/lib/linkToApp";
-import type { DemoVote } from "./demo-data";
+import type { TFunction } from "i18next";
+import { UserIcon } from "lucide-react";
+import type { DemoDay, DemoVote } from "./demo-data";
+import { formatDemoParts } from "./demo-data";
 import { DemoFrame, DemoScreen } from "./demo-frame";
+import { VoteButton } from "./vote-button";
 import { VoteCount } from "./vote-count";
-import { VoteIcon } from "./vote-icon";
+import { VoteSelector } from "./vote-selector";
 
-export type MobileDemoDay = {
-  label: string;
-  options: { id: string; time: string; score: number }[];
-};
-
+// Pre-selected votes for the first few options, so the demo opens looking
+// like a poll someone has already started filling in.
 const initialVotes: (DemoVote | undefined)[] = ["yes", "yes", "ifNeedBe"];
 
-const TriState = ({
-  label,
-  value,
-  onChange,
+export const MobileDemo = ({
+  locale,
+  days,
+  scores,
+  t,
 }: {
-  label: string;
-  value?: DemoVote;
-  onChange: (vote: DemoVote) => void;
+  locale: string;
+  days: DemoDay[];
+  scores: number[];
+  t: TFunction<"home">;
 }) => {
-  const { t } = useTranslation("home");
-  const voteLabels: Record<DemoVote, string> = {
-    yes: t("heroDemoYes", { defaultValue: "Yes" }),
-    ifNeedBe: t("heroDemoIfNeedBe", { defaultValue: "If need be" }),
-    no: t("heroDemoNo", { defaultValue: "No" }),
-  };
-  return (
-    <div className="flex h-8 w-24 shrink-0 items-center rounded-lg border border-gray-200/60 bg-gray-100 p-0.5">
-      {(["yes", "ifNeedBe", "no"] as const).map((vote) => {
-        const isSelected = vote === value;
-        return (
-          <label
-            key={vote}
-            className={cn(
-              "group flex h-full flex-1 cursor-pointer items-center justify-center rounded-md has-[:focus-visible]:ring-2 has-[:focus-visible]:ring-indigo-500",
-              isSelected && "bg-white shadow-sm",
-            )}
-          >
-            <input
-              type="radio"
-              name={label}
-              checked={isSelected}
-              onChange={() => onChange(vote)}
-              aria-label={voteLabels[vote]}
-              className="sr-only"
-            />
-            <VoteIcon
-              vote={vote}
-              className={cn(
-                !isSelected &&
-                  "text-gray-400 opacity-50 transition-opacity group-hover:opacity-100",
-              )}
-            />
-          </label>
-        );
-      })}
-    </div>
-  );
-};
-
-export const MobileDemo = ({ days }: { days: MobileDemoDay[] }) => {
-  const { t } = useTranslation("home");
-  const [submitted, setSubmitted] = React.useState(false);
-  const topScore = Math.max(
-    ...days.flatMap((day) => day.options.map((option) => option.score)),
-  );
-  const [votes, setVotes] = React.useState<
-    Record<string, DemoVote | undefined>
-  >(() => {
-    const entries: Record<string, DemoVote | undefined> = {};
-    days
-      .flatMap((day) => day.options)
-      .forEach((option, index) => {
-        entries[option.id] = initialVotes[index];
-      });
-    return entries;
-  });
+  const format = formatDemoParts(locale);
+  const topScore = Math.max(...scores);
+  let optionIndex = -1;
 
   return (
     <DemoFrame className="rounded-[1.8rem]">
@@ -95,153 +34,57 @@ export const MobileDemo = ({ days }: { days: MobileDemoDay[] }) => {
             <span className="flex size-6 items-center justify-center rounded-full bg-gray-100">
               <UserIcon className="size-3.5 text-gray-500" />
             </span>
-            <Trans ns="home" i18nKey="heroDemoYou" defaults="You" />
+            {t("heroDemoYou", { ns: "home", defaultValue: "You" })}
           </span>
         </div>
         <div className="min-h-0 flex-1 overflow-hidden">
-          {days.map((day) => (
-            <div key={day.label}>
-              <div className="border-gray-100 border-b bg-gray-50/60 px-3 py-2 font-semibold text-gray-900 text-xs">
-                {day.label}
-              </div>
-              {day.options.map((option) => (
-                <div
-                  key={option.id}
-                  className="flex items-center justify-between gap-2 border-gray-100 border-b px-3 py-3"
-                >
-                  <div className="min-w-0 flex-1 text-gray-800 text-xs">
-                    {option.time}
-                  </div>
-                  <VoteCount
-                    count={option.score}
-                    highlight={option.score === topScore}
-                    className="mr-2 shrink-0"
-                  />
-                  <TriState
-                    label={`${day.label} ${option.time}`}
-                    value={votes[option.id]}
-                    onChange={(vote) => {
-                      posthog?.capture("landing:hero_demo_vote_change", {
-                        vote,
-                      });
-                      setVotes((current) => ({
-                        ...current,
-                        [option.id]: vote,
-                      }));
-                    }}
-                  />
+          {days.map((day) => {
+            const label = format.fullDay.format(day.date);
+            return (
+              <div key={label}>
+                <div className="border-gray-100 border-b bg-gray-50/60 px-3 py-2 font-semibold text-gray-900 text-xs">
+                  {label}
                 </div>
-              ))}
-            </div>
-          ))}
+                {day.options.map((option) => {
+                  optionIndex += 1;
+                  const time = format.time.formatRange(
+                    option.start,
+                    option.end,
+                  );
+                  const score = scores[optionIndex];
+                  return (
+                    <div
+                      key={option.start.toISOString()}
+                      className="flex items-center justify-between gap-2 border-gray-100 border-b px-3 py-3"
+                    >
+                      <div className="min-w-0 flex-1 text-gray-800 text-xs">
+                        {time}
+                      </div>
+                      <VoteCount
+                        count={score}
+                        highlight={score === topScore}
+                        className="mr-2 shrink-0"
+                      />
+                      <VoteSelector
+                        label={`${label} ${time}`}
+                        defaultValue={initialVotes[optionIndex]}
+                      />
+                    </div>
+                  );
+                })}
+              </div>
+            );
+          })}
         </div>
         <div className="absolute inset-x-0 bottom-0 flex items-center gap-2 bg-gradient-to-t from-55% from-white via-85% via-white/70 to-transparent p-3 pt-10">
           <span
             aria-hidden="true"
             className="flex-1 rounded-xl border border-white/60 bg-white/70 py-2.5 text-center font-medium text-gray-800 text-sm shadow-sm backdrop-blur-md"
           >
-            <Trans ns="home" i18nKey="heroDemoDecline" defaults="Decline" />
+            {t("heroDemoDecline", { ns: "home", defaultValue: "Decline" })}
           </span>
-          <button
-            type="button"
-            onClick={() => {
-              posthog?.capture("landing:hero_demo_continue_click");
-              setSubmitted(true);
-            }}
-            className="flex-[2] cursor-pointer rounded-xl bg-indigo-500/90 py-2.5 text-center font-medium text-sm text-white shadow-sm backdrop-blur-md hover:bg-indigo-500"
-          >
-            <Trans ns="home" i18nKey="heroDemoVote" defaults="Vote" />
-          </button>
+          <VoteButton />
         </div>
-        {submitted && (
-          <div className="absolute inset-0 z-10 flex items-center justify-center bg-gray-900/20 p-4">
-            <m.div
-              initial={{ scale: 0.92, opacity: 0, y: 8 }}
-              animate={{ scale: 1, opacity: 1, y: 0 }}
-              transition={{ type: "spring", duration: 0.5, bounce: 0.35 }}
-              className="relative flex aspect-[4/3] w-full flex-col items-center justify-center rounded-2xl border border-white/60 bg-white/85 p-4 text-center shadow-lg backdrop-blur-xl"
-            >
-              <button
-                type="button"
-                aria-label={t("heroDemoClose", { defaultValue: "Close" })}
-                onClick={() => {
-                  posthog?.capture("landing:hero_demo_close_click");
-                  setSubmitted(false);
-                }}
-                className="absolute top-3 right-3 flex size-6 cursor-pointer items-center justify-center rounded-full text-gray-400 hover:bg-gray-100 hover:text-gray-600"
-              >
-                <XIcon className="size-4" />
-              </button>
-              <m.div
-                initial={{ scale: 0, opacity: 0 }}
-                animate={{ scale: 1, opacity: 1 }}
-                transition={{
-                  type: "spring",
-                  duration: 0.6,
-                  bounce: 0.45,
-                  delay: 0.1,
-                }}
-                className="relative mb-3 inline-block"
-              >
-                <div className="absolute top-0 right-0 bottom-2 -left-1.5 origin-bottom -rotate-12 scale-95 rounded-xl bg-white opacity-75 shadow-xs ring-1 ring-gray-200 ring-inset" />
-                <div className="absolute top-0 -right-1.5 bottom-2 left-0 origin-bottom rotate-12 scale-95 rounded-xl bg-white opacity-75 shadow-xs ring-1 ring-gray-200 ring-inset" />
-                <div className="relative inline-flex rounded-xl bg-white p-2.5 shadow-xs ring-1 ring-gray-200 ring-inset">
-                  <CircleCheckIcon className="size-5 text-green-500" />
-                </div>
-              </m.div>
-              <m.div
-                initial={{ opacity: 0, y: 6 }}
-                animate={{ opacity: 1, y: 0 }}
-                transition={{
-                  type: "spring",
-                  duration: 0.6,
-                  bounce: 0.3,
-                  delay: 0.2,
-                }}
-              >
-                <p className="font-semibold text-gray-900 text-sm">
-                  <Trans
-                    ns="home"
-                    i18nKey="heroDemoThatWasEasy"
-                    defaults="That was easy!"
-                  />
-                </p>
-                <p className="mt-1.5 text-gray-500 text-xs">
-                  <Trans
-                    ns="home"
-                    i18nKey="heroDemoEasyDescription"
-                    defaults="That's all it takes to respond to a poll."
-                  />
-                </p>
-              </m.div>
-              <m.div
-                initial={{ opacity: 0, y: 6 }}
-                animate={{ opacity: 1, y: 0 }}
-                transition={{
-                  type: "spring",
-                  duration: 0.6,
-                  bounce: 0.3,
-                  delay: 0.3,
-                }}
-                className="mt-4"
-              >
-                <Link
-                  href={linkToApp("/new")}
-                  onClick={() => {
-                    posthog?.capture("landing:hero_demo_modal_cta_click");
-                  }}
-                  className="font-medium text-indigo-600 text-xs hover:underline"
-                >
-                  <Trans
-                    ns="home"
-                    i18nKey="heroDemoCreatePoll"
-                    defaults="Create your own poll"
-                  />
-                </Link>
-              </m.div>
-            </m.div>
-          </div>
-        )}
       </DemoScreen>
     </DemoFrame>
   );
